requirements.txt
# FastAPI & server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Auth & validation
pydantic>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.0.0

# HTTP & CORS
python-multipart>=0.0.6

# Payments (optional)
paystack==1.5.0


# Email (optional)
# sendgrid>=6.0.0

# Redis (optional, for Celery/cache)
redis>=5.0.0
celery[redis]>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
aiosqlite>=0.19.0

# Dev
ruff>=0.1.0
black>=23.0.0
mypy>=1.0.0
